Fix duplicate downloads and filename validity checks.
The code was incorrectly downloading files from *every* repository, not just the first one found.  This caused extra time, logging, and downloads.  Now, only the first repo that has files will be used.

The filename 'valid' check is now expanded to exclude the patterns , , , and the checksum variants of those files.